+++ assets/main.scss
@@ -2,6 +2,65 @@
 @import "./media.scss";
 @import "./fonts.scss";
 
+// Page Transition
+.transition-fade {
+  display: block;
+  position: fixed;
+  inset: 0;
+  opacity: 0;
+  background-color: var(--color-background);
+  transition: opacity .5s;
+  pointer-events: none;
+  z-index: 199;
+}
+
+html.is-animating .transition-fade{
+  opacity: 1;
+}
+
+.transition-zoom {
+    background-color: var(--color-text);
+    position: fixed;
+    top: var(--top);
+    left: var(--left);
+    height: 2vmax;
+    width: 2vmax;
+    border-left-color: black;
+    z-index: 200;
+    border-radius: 50%;
+    transform-origin: clamp(20%, var(--x), 80%) clamp(20%, var(--y), 80%);
+    transform: translate(-50%, -50%) scale(1);
+    transition: border-left-color 0.6s linear;
+    opacity: 0;
+    pointer-events: none;
+}
+
+html.is-animating .transition-zoom{
+    border-left-color: white;
+    animation: zoomleave 0.6s cubic-bezier(0.550, 0.055, 0.675, 0.190) forwards;
+}
+
+// html.is-animating .transition-zoom{
+//     border-left-color: red;
+//     animation: zoom-enter 5s linear//cubic-bezier(0.550, 0.055, 0.675, 0.190);
+// }
+
+@keyframes zoomleave {
+  2%  {
+    opacity: 1;
+  }
+  80%  {
+    transform: translate(-50%, -50%) scale(200);
+    background-color: var(--color-text);
+    opacity: 1;
+  }
+  100% {
+    transform: translate(-50%, -50%) scale(200);
+    background-color: var(--color-background);
+    opacity: 1;
+  }
+}
+
 // COLORS
 :root {
   --color-background: #fff;
@@ -28,22 +87,12 @@ body {
   // overflow-x: hidden;
 }
 
-.animate-dark-mode {
-  body,
-  p,
-  img,
-  a,
-  div,
-  h1,
-  h2,
-  h3,
-  h4,
-  main,
-  section {
-    transition: background-color .2s ease,
-                color .2s ease;
-  }
-}
+// .toggle-darkmode {
+//   * {
+//     transition: background-color .2s ease,
+//                 color .2s ease;
+//   }
+// }
 
 // TYPO
 
@@ -82,7 +131,13 @@ body {
   width: 100%;
 }
 
-main.main {
+main {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+div.main {
   z-index: 10;
   position: relative;
   box-sizing: border-box;
@@ -191,7 +246,7 @@ header.header {
   filter: invert(1);
 }
 
-body.animate-dark-mode .darkmode-switch-icon {
+body.toggle-darkmode .darkmode-switch-icon {
   transition: transform .2s cubic-bezier(.42,.42,.84,1.3),
               background-image .1s linear .1s;
 }
@@ -333,13 +388,11 @@ footer.footer {
     --translateY: 0%;
     transform: translate(var(--translateX), var(--translateY));
   }
-  &:hover button {
-    transform: translate(var(--translateX), var(--translateY));
-  }
 }
 
 // HOME PAGE
 section.intro{
+  // display: none;
   background-color: var(--color-background);
   padding: 2.75rem 0;
   box-sizing: border-box;
@@ -462,7 +515,8 @@ section.work-next-section {
       color: var(--color-text);
       position: relative;
       img {
-        width: 100%;
+        max-width: 100%;
+        aspect-ratio: 650 / 384;
         object-fit: cover;
       }
       h3 {
