+++ layouts/partials/images/full.html
@@ -1,6 +1,6 @@
 <img 
     srcset="
-    {{ site.Params.cloudinaryBaseURL }}/w_360,f_auto,q_auto,fl_progressive/{{ .image }} 360w
+    {{ site.Params.cloudinaryBaseURL }}/w_360,f_auto,q_auto,fl_progressive/{{ .image }} 360w,
     {{ site.Params.cloudinaryBaseURL }}/w_500,f_auto,q_auto,fl_progressive/{{ .image }} 500w,
     {{ site.Params.cloudinaryBaseURL }}/w_720,f_auto,q_auto,fl_progressive/{{ .image }} 720w,
     {{ site.Params.cloudinaryBaseURL }}/w_1000,f_auto,q_auto,fl_progressive/{{ .image }} 1000w,
