+++ layouts/partials/modules/works.html
@@ -4,13 +4,13 @@
     </div>
     <div class="works-grid relative">
         {{ range where site.RegularPages "Type" "==" "works"}}
-            <a class="works-link" href="{{ .RelPermalink }}">
+            <a class="works-link" href="{{ .RelPermalink }}" >
                 {{ $file := (replaceRE `(.*\/v\d*\/)` "" .Params.image) }}
                 {{ partial "images/half-dektop" (dict "image" $file "class" "works-image" "alt" "{{ .Title }} preview image" "width" "650" "height" "384" "loading" "lazy") }}
                 <h3 class="works-title">{{ .Title }}</h3>
                 <span class="works-type">{{ .Param "project-type" }}</span>
                 <div class=" hover-invert-button-trigger">
-                    <button class="hover-invert-button">{{ $.button }}</button>
+                        <button class="hover-invert-button">{{ $.button }}</button>
                 </div>
             </a>
         {{ end }}
