+++ package.json
@@ -1,6 +1,10 @@
 {
   "dependencies": {
+    "@swup/body-class-plugin": "^1.0.2",
+    "@swup/debug-plugin": "^2.0.0",
+    "@swup/scroll-plugin": "^1.3.1",
     "cloudinary-core": "^2.13.0",
-    "netlify-cli": "^12.0.4"
+    "netlify-cli": "^12.0.4",
+    "swup": "^2.0.19"
   }
 }
